fix: v1で切り分け＆errorResponseを返り値に修正
